# assorted optimizations