--- a/hyperloglog.go
+++ b/hyperloglog.go
@@ -18,7 +18,7 @@ type Sketch struct {
 	p          uint8
 	m          uint32
 	alpha      float64
-	tmpSet     *set
+	tmpSet     set
 	sparseList *compressedList
 	regs       []uint8
 }
@@ -54,7 +54,7 @@ func NewSketch(precision uint8, sparse bool) (*Sketch, error) {
 		alpha: alpha(float64(m)),
 	}
 	if sparse {
-		s.tmpSet = newSet(0)
+		s.tmpSet = makeSet(0)
 		s.sparseList = newCompressedList(0)
 	} else {
 		s.regs = make([]uint8, m)
@@ -140,7 +140,7 @@ func (sk *Sketch) toNormal() {
 		sk.insert(i, r)
 	}
 
-	sk.tmpSet = nil
+	sk.tmpSet = nilSet
 	sk.sparseList = nil
 }
 
@@ -211,7 +211,7 @@ func (sk *Sketch) mergeSparse() {
 	}
 
 	sk.sparseList = newList
-	sk.tmpSet = newSet(0)
+	sk.tmpSet = makeSet(0)
 }
 
 // MarshalBinary implements the encoding.BinaryMarshaler interface.
@@ -305,7 +305,7 @@ func (sk *Sketch) UnmarshalBinary(data []byte) error {
 
 		// Unmarshal the tmp_set.
 		tssz := binary.BigEndian.Uint32(data[4:8])
-		sk.tmpSet = newSet(int(tssz))
+		sk.tmpSet = makeSet(int(tssz))
 
 		// We need to unmarshal tssz values in total, and each value requires us
 		// to read 4 bytes.
@@ -321,7 +321,7 @@ func (sk *Sketch) UnmarshalBinary(data []byte) error {
 
 	// Using the dense Sketch.
 	sk.sparseList = nil
-	sk.tmpSet = nil
+	sk.tmpSet = nilSet
 
 	if v == 1 {
 		return sk.unmarshalBinaryV1(data[8:], b)