--- a/sparse.go
+++ b/sparse.go
@@ -108,9 +108,3 @@ func (s *set) AppendBinary(data []byte) ([]byte, error) {
 
 	return data, nil
 }
-
-type uint64Slice []uint32
-
-func (p uint64Slice) Len() int           { return len(p) }
-func (p uint64Slice) Less(i, j int) bool { return p[i] < p[j] }
-func (p uint64Slice) Swap(i, j int)      { p[i], p[j] = p[j], p[i] }