See the commit messages for details and `Benchmark_Add` comparisons between each commit.

Before/after of `Benchmark_Add` with all commits:

```
name               old time/op    new time/op    delta
_Add_100-10          1.20µs ± 0%    1.08µs ± 0%   -9.59%  (p=0.000 n=9+10)
_Add_1000-10          123µs ± 1%      90µs ± 0%  -26.78%  (p=0.000 n=10+9)
_Add_10000-10        2.56ms ± 0%    1.76ms ± 0%  -31.19%  (p=0.000 n=9+9)
_Add_100000-10        630µs ± 0%     627µs ± 0%   -0.43%  (p=0.000 n=9+10)
_Add_1000000-10      6.36ms ± 1%    6.27ms ± 0%   -1.30%  (p=0.000 n=10+9)
_Add_10000000-10     63.6ms ± 1%    62.9ms ± 0%   -1.12%  (p=0.001 n=8+9)
_Add_100000000-10     636ms ± 1%     628ms ± 0%   -1.22%  (p=0.000 n=9+9)

name               old alloc/op   new alloc/op   delta
_Add_100-10           0.00B          0.00B          ~     (all equal)
_Add_1000-10         38.6kB ± 0%    38.5kB ± 0%   -0.12%  (p=0.000 n=9+9)
_Add_10000-10         666kB ± 0%     666kB ± 0%   -0.06%  (p=0.000 n=10+10)
_Add_100000-10       1.92kB ± 0%    1.80kB ± 0%   -6.22%  (p=0.000 n=7+10)
_Add_1000000-10      16.7kB ± 2%    16.3kB ± 0%   -2.43%  (p=0.000 n=10+9)
_Add_10000000-10      171kB ± 0%     192kB ± 0%  +12.44%  (p=0.000 n=8+10)
_Add_100000000-10    1.54MB ± 0%    1.54MB ± 0%   -0.05%  (p=0.000 n=10+10)

name               old allocs/op  new allocs/op  delta
_Add_100-10            0.00           0.00          ~     (all equal)
_Add_1000-10           24.0 ± 0%      21.0 ± 0%  -12.50%  (p=0.000 n=10+10)
_Add_10000-10           243 ± 0%       213 ± 0%  -12.35%  (p=0.000 n=10+10)
_Add_100000-10         0.00           0.00          ~     (all equal)
_Add_1000000-10        4.00 ± 0%      3.00 ± 0%  -25.00%  (p=0.000 n=10+9)
_Add_10000000-10       46.0 ± 0%      45.0 ± 0%   -2.17%  (p=0.002 n=8+10)
_Add_100000000-10       416 ± 0%       364 ± 0%  -12.50%  (p=0.000 n=10+10)
```

Before/after of `Benchmark_Merge` with all commits:

```
name                                   old time/op    new time/op    delta
_Merge/size1=100/size2=100-10            9.62µs ± 1%    6.66µs ± 1%  -30.73%  (p=0.000 n=8+9)
_Merge/size1=100/size2=10000-10          54.8µs ± 1%    51.8µs ± 1%   -5.53%  (p=0.000 n=10+8)
_Merge/size1=100/size2=1000000-10        15.9µs ± 0%    14.5µs ± 1%   -8.66%  (p=0.000 n=8+9)
_Merge/size1=10000/size2=100-10          51.1µs ± 5%    50.4µs ± 2%     ~     (p=0.315 n=10+10)
_Merge/size1=10000/size2=10000-10         101µs ± 3%     101µs ± 4%     ~     (p=0.579 n=10+10)
_Merge/size1=10000/size2=1000000-10      61.8µs ± 2%    61.6µs ± 2%     ~     (p=0.579 n=10+10)
_Merge/size1=1000000/size2=100-10        11.1µs ± 1%    11.0µs ± 2%     ~     (p=0.063 n=10+10)
_Merge/size1=1000000/size2=10000-10      19.7µs ± 2%    19.5µs ± 0%   -1.31%  (p=0.000 n=9+10)
_Merge/size1=1000000/size2=1000000-10    56.8µs ± 3%    56.0µs ± 2%   -1.34%  (p=0.017 n=9+10)

name                                   old alloc/op   new alloc/op   delta
_Merge/size1=100/size2=100-10            10.2kB ± 0%    10.2kB ± 0%   -0.39%  (p=0.000 n=10+10)
_Merge/size1=100/size2=10000-10          21.6kB ± 0%    21.5kB ± 0%   -0.19%  (p=0.000 n=10+10)
_Merge/size1=100/size2=1000000-10        21.6kB ± 0%    21.5kB ± 0%   -0.19%  (p=0.000 n=10+10)
_Merge/size1=10000/size2=100-10          16.6kB ± 0%    16.5kB ± 0%   -0.05%  (p=0.000 n=10+10)
_Merge/size1=10000/size2=10000-10        16.6kB ± 0%    16.5kB ± 0%   -0.05%  (p=0.000 n=10+10)
_Merge/size1=10000/size2=1000000-10      16.6kB ± 0%    16.5kB ± 0%   -0.05%  (p=0.000 n=10+10)
_Merge/size1=1000000/size2=100-10        16.6kB ± 0%    16.5kB ± 0%   -0.05%  (p=0.000 n=10+10)
_Merge/size1=1000000/size2=10000-10      16.6kB ± 0%    16.5kB ± 0%   -0.05%  (p=0.000 n=10+10)
_Merge/size1=1000000/size2=1000000-10    16.6kB ± 0%    16.5kB ± 0%   -0.05%  (p=0.000 n=10+10)

name                                   old allocs/op  new allocs/op  delta
_Merge/size1=100/size2=100-10              20.0 ± 0%      17.0 ± 0%  -15.00%  (p=0.000 n=10+10)
_Merge/size1=100/size2=10000-10            20.0 ± 0%      17.0 ± 0%  -15.00%  (p=0.000 n=10+10)
_Merge/size1=100/size2=1000000-10          20.0 ± 0%      17.0 ± 0%  -15.00%  (p=0.000 n=10+10)
_Merge/size1=10000/size2=100-10            6.00 ± 0%      5.00 ± 0%  -16.67%  (p=0.000 n=10+10)
_Merge/size1=10000/size2=10000-10          6.00 ± 0%      5.00 ± 0%  -16.67%  (p=0.000 n=10+10)
_Merge/size1=10000/size2=1000000-10        6.00 ± 0%      5.00 ± 0%  -16.67%  (p=0.000 n=10+10)
_Merge/size1=1000000/size2=100-10          6.00 ± 0%      5.00 ± 0%  -16.67%  (p=0.000 n=10+10)
_Merge/size1=1000000/size2=10000-10        6.00 ± 0%      5.00 ± 0%  -16.67%  (p=0.000 n=10+10)
_Merge/size1=1000000/size2=1000000-10      6.00 ± 0%      5.00 ± 0%  -16.67%  (p=0.000 n=10+10)
```