@lukasmalkmus 2025-02-28
Released this as part of `v0.2.4`. For the first time I used GitHub to create the tag and the release, let me know if something doesn't seem right.

https://github.com/axiomhq/hyperloglog/releases/tag/v0.2.4.